FDAB-47: Django Data Wizard package
Change base template for Django Data Wizard package so that it does not overwrite the CSS specified in its parent template. Change auto template for Django Data Wizard package so that it does not overwrite the JavaScript specified in its parent template. Change auto template for Django Data Wizard package so that it uses a local, customized and compressed version of progress.js and so that it excludes a redundant reference to the JQuery library that is already included via the base template. Add call to update session expiry in local, customized and compressed version of progress.js. Change spinner image indicating AJAX is running on the client from red to grey.